[DEVMAJOR] Add "+customX" syntax to version strings
This replaces the '0.0.0-anything' trick to allow test versions based
on proper versions: a "1.2.3-rc4+feature1" version is newer than a
normal "1.2.3-rc4", but older than a "1.2.3-rc5" version.

The "+customX" part is composed of an up-to-32 characters([a-z_-])
name, and a non-zero number. e.g. "+test1" is valid, while "+1" and
"+test", or "+test0" are not.

puf_compare_versions returns 0 (same version) for versions with the
same root, but different custom parts, even if they only differ by
custom number.

This change also adds unit testing for puf_version_xxx functions.