[DOC] Fix documentation for puf_walk()
The documentation for puf_walk() states that it returns either 0 on
success or a negative errno on failure. That description is incorrect.
Actually, puf_walk() may return:
- a negative errno originating from libpuf code itself (e.g. in case of
  a corrupted archive)
- a negative value returned by the user in one of the callbacks to
  signal an error
- the value 0, when the user interrupts the walk by returning 0 from one
  of the callbacks
- the last positive value returned by one of the callbacks when
  puf_walk() reaches the end of file

I don't think that last case is a bug, since it can be useful to know
whether puf_walk() was stopped early or ran until the end of file.
So it means the documentation is wrong and this commit attempts to fix
it.